[Docs] Documentation update
A few warning fixes

+++ src/ast/astgen/misc.c
@@ -10,7 +10,7 @@ int var_lookup(ast_node_t* node, ast_ctx_t* ctx, sym_table_t* smt) {
     if (TKN_isvariable(node->t)) {
         variable_info_t varinfo = { .type = UNKNOWN_NUMERIC_TOKEN };
         for (int s = ctx->scopes.stack.top; s >= 0; s--) {
-            short s_id = (short)ctx->scopes.stack.data[s].d;
+            short s_id = (short)((long)ctx->scopes.stack.data[s].d);
             if (VRTB_get_info(node->t->body, s_id, &varinfo, &smt->v)) {
                 node->sinfo.v_id    = varinfo.v_id;
                 node->sinfo.s_id    = varinfo.s_id;
